Reorganize test utilities to prevent implicit TestSupport product
