+++ Tests/DeferredTests/FutureIgnoreTests.swift
@@ -8,9 +8,6 @@
 
 import XCTest
 @testable import Deferred
-#if SWIFT_PACKAGE
-@testable import TestSupport
-#endif
 
 class FutureIgnoreTests: XCTestCase {
     static var allTests: [(String, (FutureIgnoreTests) -> () throws -> Void)] {
