+++ Tests/TaskTests/TaskComprehensiveTests.swift
@@ -17,7 +17,6 @@ import Dispatch
 #if SWIFT_PACKAGE
     import Deferred
     @testable import Task
-    @testable import TestSupport
 #else
     @testable import Deferred
 #endif
