+++ Tests/TaskTests/TaskGroupTests.swift
@@ -12,7 +12,6 @@ import Dispatch
 #if SWIFT_PACKAGE
 import Deferred
 @testable import Task
-@testable import TestSupport
 #else
 @testable import Deferred
 #endif
